2024AppControl/src/main/java/br/edu/infnet/AppControl/model/domain/Aluno.java
package br.edu.infnet.AppControl.model.domain;

public class Aluno {
	
	private int id;
	private String nome;
	private String tipo;
	private int frequencia;
	private float nota;
	private boolean aprovado;
	

	public boolean status(){
		return nota>7?true:false;
	}
	
	public int getFrequencia() {
		return frequencia;
	}

	public void setFrequencia(int frequencia) {
		this.frequencia = frequencia;
	}

	public boolean isAprovado() {
		return aprovado;
	}

	public void setAprovado(boolean aprovado) {
		this.aprovado = aprovado;
	}	
	
	public float getNota() {
		return nota;
	}
	public void setNota(float nota) {
		this.nota = nota;
	}
	
	public int getId() {
		return id;
	}
	public void setId(int id) {
		this.id = id;
	}
	public String getNome() {
		return nome;
	}
	public void setNome(String nome) {
		this.nome = nome;
	}
	public String getTipo() {
		return tipo;
	}
	public void setTipo(String tipo) {
		this.tipo = tipo;
	}
	
	

}
